Move voucher code to separate app.
