+++ apps/fund/urlsapi.py
@@ -1,8 +1,7 @@
 from django.conf.urls import patterns, include, url
 from surlex.dj import surl
-from .views import OrderList, OrderDetail, OrderCurrentDetail, PaymentProfileCurrent,  VoucherDetail, \
-    OrderVoucherList, OrderVoucherDetail, VoucherDonationList, VoucherDonationDetail,  CustomVoucherRequestList, \
-    PaymentCurrent, RecurringDirectDebitPaymentList, RecurringDirectDebitPaymentDetail, OrderCurrentDonationDetail, \
+from .views import OrderList, OrderDetail, OrderCurrentDetail, PaymentProfileCurrent,  PaymentCurrent, \
+    RecurringDirectDebitPaymentList, RecurringDirectDebitPaymentDetail, OrderCurrentDonationDetail, \
     OrderCurrentDonationList, NestedDonationDetail, NestedDonationList, DonationList, DonationDetail, TickerList
 
 
@@ -12,9 +11,6 @@ urlpatterns = patterns('',
     surl(r'^orders/<pk:#>$', OrderDetail.as_view(), name='fund-order-detail'),
     surl(r'^orders/<order_pk:#>/donations/$', NestedDonationList.as_view(), name='fund-order-donation-list'),
     surl(r'^orders/<order_pk:#>/donations/<pk:#>$', NestedDonationDetail.as_view(), name='fund-order-donation-detail'),
-    # Voucher code is disabled for now.
-    # surl(r'^orders/<order_pk:#>/vouchers/$', OrderVoucherList.as_view(), name='fund-order-voucher-list'),
-    # surl(r'^orders/<order_pk:#>/vouchers/<pk:#>$', OrderVoucherDetail.as_view(), name='fund-order-voucher-detail'),
 
     # Donations
     surl(r'^donations/$', DonationList.as_view(), name='fund-donation-list'),
@@ -27,15 +23,6 @@ urlpatterns = patterns('',
     url(r'^orders/current$', OrderCurrentDetail.as_view(), {'alias': 'current'}, name='fund-order-current-detail'),
     url(r'^orders/current/donations/$', OrderCurrentDonationList.as_view(), {'alias': 'current'}, name='fund-order-current-donation-list'),
     surl(r'^orders/current/donations/<pk:#>$', OrderCurrentDonationDetail.as_view(), {'alias': 'current'}, name='fund-order-current-donation-detail'),
-    # Voucher code is disabled for now.
-    # url(r'^orders/current/vouchers/$', OrderVoucherList.as_view(), {'alias': 'current'}, name='fund-order-current-voucher-list'),
-    # surl(r'^orders/current/vouchers/<pk:#>$', OrderVoucherDetail.as_view(), {'alias': 'current'}, name='fund-order-current-voucher-detail'),
-    #
-    # # Vouchers
-    # surl(r'^vouchers/<code:s>$', VoucherDetail.as_view(), name='voucher-detail'),
-    # surl(r'^vouchers/<code:s>/donations/$', VoucherDonationList.as_view(), name='voucher-donation-list'),
-    # surl(r'^vouchers/<code:s>/donations/<pk:#>$', VoucherDonationDetail.as_view(), name='voucher-donation-list'),
-    # surl(r'^customvouchers/$', CustomVoucherRequestList.as_view(), name='custom-voucher-request-list'),
 
     # Payments
     url(r'^paymentprofiles/current$', PaymentProfileCurrent.as_view(), name='payment-profile-current'),
@@ -44,5 +31,4 @@ urlpatterns = patterns('',
     url(r'', include('apps.cowry.urlsapi')),
     url(r'^recurringdirectdebitpayments/$', RecurringDirectDebitPaymentList.as_view(), name='recurring-direct-debit-payment-list'),
     surl(r'^recurringdirectdebitpayments/<pk:#>$', RecurringDirectDebitPaymentDetail.as_view(), name='recurring-direct-debit-payment-detail'),
-
 )
