+++ apps/fund/views.py
@@ -11,17 +11,13 @@ from apps.fund.serializers import DonationInfoSerializer, NestedDonationSerializ
 from django.contrib.auth.models import AnonymousUser
 from django.contrib.auth.signals import user_logged_in
 from registration.signals import user_registered
-from django.contrib.contenttypes.models import ContentType
 from django.db import transaction
 from django.http import Http404
-from rest_framework import mixins, exceptions, status, permissions, response, generics
+from rest_framework import exceptions, status, permissions, response, generics
 from django.utils.translation import ugettext as _
-from .mails import mail_voucher_redeemed, mail_custom_voucher_request
-from .models import Donation, Order, OrderStatuses, Voucher, CustomVoucherRequest, VoucherStatuses, \
-    DonationStatuses, RecurringDirectDebitPayment
+from .models import Donation, Order, OrderStatuses, DonationStatuses, RecurringDirectDebitPayment
 from .permissions import IsUser
-from .serializers import DonationSerializer, OrderSerializer, VoucherSerializer, VoucherDonationSerializer, \
-    VoucherRedeemSerializer, CustomVoucherRequestSerializer, RecurringDirectDebitPaymentSerializer, \
+from .serializers import DonationSerializer, OrderSerializer, RecurringDirectDebitPaymentSerializer, \
     OrderCurrentSerializer, OrderCurrentDonationSerializer
 
 
@@ -263,7 +259,6 @@ class CurrentOrderMixin(object):
         queryset = self.model.objects.filter(order=order)
         return queryset
 
-
     def create(self, request, *args, **kwargs):
         order = self.get_current_order()
         if not order:
@@ -271,12 +266,11 @@ class CurrentOrderMixin(object):
         serializer = self.get_serializer(data=request.DATA, files=request.FILES)
 
         if serializer.is_valid():
+            if request.user.is_authenticated():
+                setattr(serializer.object, self.user_field, request.user)
+            serializer.object.order = order
             self.pre_save(serializer.object)
             self.object = serializer.save(force_insert=True)
-            if request.user.is_authenticated():
-                setattr(self.object, self.user_field, request.user)
-            self.object.order = order
-            self.object.save()
             self.post_save(self.object, created=True)
 
             headers = self.get_success_headers(serializer.data)
@@ -495,129 +489,6 @@ def link_anonymous_donations(sender, user, request, **kwargs):
 
 # On account activation try to connect anonymous donations to this user.
 user_registered.connect(link_anonymous_donations)
-
-
-class OrderVoucherList(CurrentOrderMixin, generics.ListCreateAPIView):
-    """
-    Resource for ordering Vouchers
-    """
-    model = Voucher
-    serializer_class = VoucherSerializer
-    permissions_classes = (permissions.IsAuthenticatedOrReadOnly,)
-    paginate_by = 50
-    user_field = 'sender'
-
-
-class OrderVoucherDetail(CurrentOrderMixin, generics.RetrieveUpdateDestroyAPIView):
-    """
-    Resource for changing a Voucher order
-    """
-    model = Voucher
-    serializer_class = VoucherSerializer
-
-
-class VoucherMixin(object):
-
-    def get_voucher(self):
-        """
-        Override default to have semantic error responses.
-        """
-        code = self.kwargs.get('code', None)
-        if not code:
-            raise exceptions.ParseError(detail=_(u"No gift card code supplied"))
-        try:
-            voucher = Voucher.objects.get(code=code.upper())
-        except Voucher.DoesNotExist:
-            raise exceptions.ParseError(detail=_(u"No gift card with that code"))
-        if voucher.status != VoucherStatuses.paid:
-            raise exceptions.PermissionDenied(detail=_(u"Gift card code already used"))
-        return voucher
-
-
-class VoucherDetail(VoucherMixin, generics.RetrieveUpdateAPIView):
-    """
-    Resource for Voucher redemption
-    """
-    model = Voucher
-    serializer_class = VoucherRedeemSerializer
-
-    def get_object(self, queryset=None):
-        voucher = self.get_voucher()
-        self.check_object_permissions(self.request, voucher)
-        return voucher
-
-    def pre_save(self, obj):
-        mail_voucher_redeemed(obj)
-        if self.request.user.is_authenticated():
-            obj.receiver = self.request.user
-        if obj.status == VoucherStatuses.cashed:
-            for donation in obj.donations.all():
-                donation.status = DonationStatuses.paid
-                donation.save()
-
-
-class VoucherDonationList(VoucherMixin, generics.ListCreateAPIView):
-    model = Donation
-    serializer_class = VoucherDonationSerializer
-
-    def pre_save(self, obj):
-        voucher = self.get_voucher()
-        # Clear previous donations for this voucher
-        for donation in voucher.donations.all():
-            donation.delete()
-        obj.amount = voucher.amount
-
-        """
-        Keep this code around for if we want to change to multiple donations for one voucher.
-
-        count = len(voucher.donations.all()) + 1
-        rest_amount = voucher.amount
-        part_amount = floor(voucher.amount / count)
-        for donation in voucher.donations.all():
-            rest_amount -= part_amount
-            donation.amount = part_amount
-            donation.save()
-
-        obj.amount = rest_amount
-        """
-
-        obj.save()
-        voucher.donations.add(obj)
-
-    def get_queryset(self):
-        voucher = self.get_voucher()
-        return voucher.donations.all()
-
-
-# Not used yet, until we want multiple donations per voucher
-class VoucherDonationDetail(VoucherMixin, generics.RetrieveDestroyAPIView):
-    model = Donation
-    serializer_class = VoucherDonationSerializer
-
-    def destroy(self, request, *args, **kwargs):
-        obj = self.get_object()
-        voucher = self.get_voucher()
-        count = len(voucher.donations.all()) - 1
-        if count:
-            part_amount = round(100 * voucher.amount / count) / 100
-            rest_amount = voucher.amount
-            for donation in voucher.donations.all():
-                rest_amount -= part_amount
-                donation.amount = part_amount
-                donation.save()
-            # FIXME There's a bug here - donation is not defined anymore.
-            donation.amount += rest_amount
-            donation.save()
-        obj.delete()
-        return response.Response(status=status.HTTP_204_NO_CONTENT)
-
-
-class CustomVoucherRequestList(generics.ListCreateAPIView):
-    model = CustomVoucherRequest
-    serializer_class = CustomVoucherRequestSerializer
-
-    def pre_save(self, obj):
-        mail_custom_voucher_request(obj)
 
 
 # For showing the latest donations
