+++ apps/vouchers/templates/voucher_redeemed.mail.html
