+++ onepercentclub/settings/base.py
@@ -247,6 +247,7 @@ INSTALLED_APPS = (
     'apps.projects',
     'apps.fund',
     'apps.donations',
+    'apps.vouchers',
     'bluebottle.geo',
     'apps.hbtemplates',
     'apps.wallposts',
